serial-state-machine/requirements.txt
pyserial_asyncio==0.5
